split tagging feature into its own module
